fix: disable gas in gethstyle
